requirements.txt
asgiref==3.2.10
beautifulsoup4==4.10.0
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.4.1
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.1
django-bootstrap4==21.1
django-crum==0.7.9
django-import-export==2.6.1
django-pandas==0.6.6
django-rest-framework==0.1.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
et-xmlfile==1.1.0
fonttools==4.28.2
gunicorn==20.1.0
html5lib==1.1
importlib-metadata==2.1.2
kiwisolver==1.3.2
MarkupPy==1.14
matplotlib==3.5.0
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pycparser==2.21
pydyf==0.1.2
pyparsing==3.0.6
pyphen==0.11.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.3
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.1.0
tinycss2==1.1.1
tomli==1.2.2
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
zipp==3.6.0
zopfli==0.1.9
